+++ WebTools.NET.ContentAnalysis/Analysis/HtmlElementMetadata.cs
@@ -1,0 +1,9 @@
+using AngleSharp.Dom;
+
+namespace WebTools.NET.ContentAnalysis.Analysis;
+
+internal static class HtmlElementMetadata
+{
+    public static string GetClassAndIdValue(IElement element) =>
+        $"{element.GetAttribute("class")} {element.GetAttribute("id")}";
+}